+++ example/gtk/meson.build
@@ -7,13 +7,13 @@ example_gtk_inc = '../samples'
 example_gtk_deps = [
     dependency('gtkmm-3.0'),
     dependency('gtksourceviewmm-3.0'),
-    clatexmath_cairo_dep
+    clatexmath_cairomm_dep
 ]
 
 executable(
     'LaTeXGtkExample', example_gtk_src,
     include_directories: example_gtk_inc,
-    link_with: [clatexmath_cairo_lib, clatexmath_lib],
+    link_with: [],
     dependencies: example_gtk_deps,
     install: true
 )
