+++ meson.build
@@ -11,14 +11,14 @@ endif
 
 subdir('lib')
 
-if get_option('CAIRO')
-  subdir('platform/cairo')
+if get_option('CAIROMM')
+  subdir('platform/cairomm')
   if get_option('EXAMPLE_GTK')
     subdir('example/gtk')
   endif
 else
   if get_option('EXAMPLE_GTK')
-    error('EXAMPLE_GTK was enabled, but it requires CAIRO platform')
+    error('EXAMPLE_GTK was enabled, but it requires CAIROMM platform')
   endif
 endif
