+++ platform/cairomm/graphic_cairomm.h
